temporary completed prototyping

+++ talkative/SupportingFiles/UserModel.swift
@@ -31,7 +31,7 @@ class UserModel: Identifiable {
     var callCountAsLearner: Int
     var callCountAsNative: Int
     var location: GeoPoint
-    var point: Int
+    var point: Double
     var createdAt: Date // data of registration
     var updatedAt: Date // date of updating of user info
     var lastLoginBonus: Date
@@ -58,7 +58,7 @@ class UserModel: Identifiable {
         self.callCountAsLearner = from.get("callCountAsLearner") as? Int ?? 1
         self.callCountAsNative = from.get("callCountAsNaitve") as? Int ?? 1
         self.location = from.get("location") as? GeoPoint ?? GeoPoint(latitude: 0.0, longitude: 0.0)
-        self.point = from.get("point") as? Int ?? 0
+        self.point = from.get("point") as? Double ?? 0.0
         self.timestamp = from.get("createdAt") as? Timestamp ?? Timestamp()
         self.createdAt = timestamp.dateValue()
         self.timestamp = from.get("updatedAt") as? Timestamp ?? Timestamp()
@@ -111,7 +111,7 @@ class RealmUserModel: Object{
     @objc dynamic var ratingAsNative = 5.0
     @objc dynamic var callCountAsLearner = 1
     @objc dynamic var callCountAsNative = 1
-    @objc dynamic var point = 100
+    @objc dynamic var point = 100.0
     @objc dynamic var createdAt = Date() // data of registration
     @objc dynamic var updatedAt = Date() // date of updating of user info
     @objc dynamic var lastLoginBonus = Date()
