+++ talkative/SupportingFiles/pointHistoryModel.swift
@@ -14,7 +14,7 @@ struct pointHistoryModel {
     var learnerID: String
     var nativeID: String
     var imageURL: URL
-    var point: Int
+    var point: Double
     var method: Int
     var createdAt: Date
     var timestamp: Timestamp
@@ -23,7 +23,7 @@ struct pointHistoryModel {
         self.learnerID = from.get("learnerID") as? String ?? ""
         self.nativeID = from.get("nativeID") as? String ?? ""
         self.imageURL = URL(string: from.get("imageURL") as? String ?? "https://firebasestorage.googleapis.com/v0/b/talkative-21c6c.appspot.com/o/profImages%2Favatar1.jpg?alt=media&token=76d3a6c2-42bd-45f9-badb-d6398b129eaf")!
-        self.point = from.get("point") as? Int ?? 0
+        self.point = from.get("point") as? Double ?? 0.0
         self.method = from.get("method") as? Int ?? 0
         self.timestamp = from.get("createdAt") as? Timestamp ?? Timestamp()
         self.createdAt = timestamp.dateValue()
